% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dev.R
\name{check_target_colnms}
\alias{check_target_colnms}
\title{tested function for checking if all target colunms are listed in a dataframe}
\usage{
check_target_colnms(target_colnms, tab_in, ...)
}
\description{
tested function for checking if all target colunms are listed in a dataframe
}
\author{
yw
}
