% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcaproc_functions.R
\name{entropy}
\alias{entropy}
\title{Entrophy Function
This function allows you to calculate the entrophy.}
\usage{
entropy(p)
}
\arguments{
\item{p}{p value .}
}
\description{
Entrophy Function
This function allows you to calculate the entrophy.
}
\examples{
entropy()
}
\keyword{cats}
