% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dev.R
\name{file_path_list}
\alias{file_path_list}
\title{function to make a list of filenames in full path}
\usage{
file_path_list(
  keys = NULL,
  out_dir = NULL,
  prefix = NULL,
  postfix = NULL,
  sep = "_",
  extension = "csv",
  ...
)
}
\arguments{
\item{keys}{a list of key strings for major part of file names to be created}

\item{out_dir}{output dictonary}

\item{prefix}{prefix}

\item{postfix}{postfix}

\item{extension}{extension}
}
\description{
function to make a list of filenames in full path
}
\author{
yw
}
