% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcaproc_functions.R
\name{make_lca_res}
\alias{make_lca_res}
\title{Make result Function}
\usage{
make_lca_res(
  f = NULL,
  dat = NULL,
  seed = 1,
  maxiter = 1000,
  nrep = 1,
  verbose = F,
  mod_id = "Mod_0_",
  ...
)
}
\description{
Make result Function
}
