% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcaproc_functions.R
\name{summarize_stats}
\alias{summarize_stats}
\title{Make Summary of statistics from lm regression output}
\usage{
summarize_stats(mod_tmp)
}
\description{
Make Summary of statistics from lm regression output
}
